fix model issues
