+++ Projects/Forever/Sources/AnimatedMoneyLabel.swift
@@ -3,6 +3,7 @@ import Form
 import Foundation
 import UIKit
 import hCore
+import hGraphQL
 
 struct AnimatedSavingsLabel {
     let from: ReadSignal<MonetaryAmount?>
