+++ Projects/Forever/Sources/EmptyStateHeader.swift
@@ -4,6 +4,7 @@ import Foundation
 import UIKit
 import hCore
 import hCoreUI
+import hGraphQL
 
 struct EmptyStateHeader {
     let isHiddenSignal = ReadWriteSignal<Bool>(true)
